docs: add a license
Rationale: BSL or MIT would be preferred, but this library can be
considered as a derivative work and to avoid any possible complications
I have adopted BSD-3-Clause as the original work and just added my name
and institution (hoping that this doesn't cause other complications).

Line breaks in the LICENSE file were tweaked in such a way that GitHub
(i.e., the Licensee gem) can detect the license type.